refactor: move the MarketParticipantList
