+++ components/common/MarketParticipant/index.tsx
@@ -3,13 +3,13 @@ import { CSSProperties, FC, useEffect, useState } from 'react';
 import { classNames } from '@/utils/index';
 import { fadeInDown } from '@/utils/animations';
 import { RoleId } from '@/types/roles';
-import HammerIcon from '../icons/HammerIcon';
-import PoundcashTag from '../icons/PoundcashTag';
-import CartPlus from '../icons/CartPlus';
-import OfferIcon from '../icons/OfferIcon';
-import ProjectBiodiversity from './ProjectBiodiversity';
-import ProjectNutrients from './ProjectNutrients';
-import { MarketParticipantTitle } from '../../common/MarketParticipantTitle';
+import HammerIcon from '../../walkthroughs/icons/HammerIcon';
+import PoundcashTag from '../../walkthroughs/icons/PoundcashTag';
+import CartPlus from '../../walkthroughs/icons/CartPlus';
+import OfferIcon from '../../walkthroughs/icons/OfferIcon';
+import ProjectBiodiversity from '../../walkthroughs/Main/ProjectBiodiversity';
+import ProjectNutrients from '../../walkthroughs/Main/ProjectNutrients';
+import { MarketParticipantTitle } from '../MarketParticipantTitle';
 import { Products } from '../../../types/products';
 
 const PROJECT_HEIGHT = 120;
